results/requirements.txt
Flask==1.1.1
mysql-connector-python

